updated output_csv and amr_monitor
made output csv output information to a new file every time and add that same information to the database. Still needs to check database for information and only works with lists of dictionaries, but is a start. Started moving all code to the amr_monitor file so that it is central