several updates
utils.py
added a check_text option to ensure that the web page was able to retrieve text.
made use of fuzzywuzzy to extract information from ChatGPT's response in the event that it changed the names of specified variables slightly.

install_required_modules.bat
added new modules

chatgpt_api.py
added token limit

request.txt
added some more text as the ChatGPT was chosing to not respond at times. This was to convince it to respond. Results have been questionable.